[jdbc] rename record packages
db.sql -> db.schema
db.sql.statement -> db.sql
