+++ jdbc/db/core/src/main/java/org/eclipse/daanse/common/jdbc/db/core/DatabaseServiceImpl.java
@@ -47,12 +47,12 @@ import org.eclipse.daanse.common.jdbc.db.record.meta.MetaInfoR;
 import org.eclipse.daanse.common.jdbc.db.record.meta.TableDefinitionR;
 import org.eclipse.daanse.common.jdbc.db.record.meta.TableMetaDataR;
 import org.eclipse.daanse.common.jdbc.db.record.meta.TypeInfoR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.CatalogReferenceR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.ColumnDefinitionR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.ColumnMetaDataR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.ColumnReferenceR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.SchemaReferenceR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.TableReferenceR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.CatalogReferenceR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.ColumnDefinitionR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.ColumnMetaDataR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.ColumnReferenceR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.SchemaReferenceR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.TableReferenceR;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.ServiceScope;
 import org.slf4j.Logger;
