+++ jdbc/db/core/src/test/java/org/eclipse/daanse/common/jdbc/db/core/SqlStatementGeneratorImplTest.java
@@ -25,17 +25,17 @@ import org.eclipse.daanse.common.jdbc.db.api.meta.TypeInfo;
 import org.eclipse.daanse.common.jdbc.db.record.meta.DatabaseInfoR;
 import org.eclipse.daanse.common.jdbc.db.record.meta.IdentifierInfoR;
 import org.eclipse.daanse.common.jdbc.db.record.meta.MetaInfoR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.ColumnDefinitionR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.ColumnMetaDataR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.ColumnReferenceR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.SchemaReferenceR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.TableReferenceR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.statement.CreateContainerSqlStatementR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.statement.CreateSchemaSqlStatementR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.statement.DropContainerSqlStatementR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.statement.DropSchemaSqlStatementR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.statement.InsertSqlStatementR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.statement.TruncateTableSqlStatementR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.ColumnDefinitionR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.ColumnMetaDataR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.ColumnReferenceR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.SchemaReferenceR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.TableReferenceR;
+import org.eclipse.daanse.common.jdbc.db.record.sql.CreateContainerSqlStatementR;
+import org.eclipse.daanse.common.jdbc.db.record.sql.CreateSchemaSqlStatementR;
+import org.eclipse.daanse.common.jdbc.db.record.sql.DropContainerSqlStatementR;
+import org.eclipse.daanse.common.jdbc.db.record.sql.DropSchemaSqlStatementR;
+import org.eclipse.daanse.common.jdbc.db.record.sql.InsertSqlStatementR;
+import org.eclipse.daanse.common.jdbc.db.record.sql.TruncateTableSqlStatementR;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
