+++ jdbc/db/record/src/main/java/org/eclipse/daanse/common/jdbc/db/record/schema/ColumnReferenceR.java
@@ -11,12 +11,16 @@
 *   SmartCity Jena - initial
 *   Stefan Bischof (bipolis.org) - initial
 */
-package org.eclipse.daanse.common.jdbc.db.record.sql.element;
+package org.eclipse.daanse.common.jdbc.db.record.schema;
 
-import org.eclipse.daanse.common.jdbc.db.api.schema.ColumnDefinition;
-import org.eclipse.daanse.common.jdbc.db.api.schema.ColumnMetaData;
+import java.util.Optional;
+
 import org.eclipse.daanse.common.jdbc.db.api.schema.ColumnReference;
+import org.eclipse.daanse.common.jdbc.db.api.schema.TableReference;
 
-public record ColumnDefinitionR(ColumnReference column, ColumnMetaData columnType) implements ColumnDefinition {
+public record ColumnReferenceR(Optional<TableReference> table, String name) implements ColumnReference {
 
+    public ColumnReferenceR(String name) {
+        this(Optional.empty(), name);
+    }
 }
