+++ jdbc/db/record/src/main/java/org/eclipse/daanse/common/jdbc/db/record/schema/package-info.java
@@ -11,12 +11,7 @@
 *   SmartCity Jena - initial
 *   Stefan Bischof (bipolis.org) - initial
 */
-package org.eclipse.daanse.common.jdbc.db.record.sql.statement;
+@org.osgi.annotation.bundle.Export
+@org.osgi.annotation.versioning.Version("0.0.1")
 
-import org.eclipse.daanse.common.jdbc.db.api.schema.SchemaReference;
-import org.eclipse.daanse.common.jdbc.db.api.sql.CreateSchemaSqlStatement;
-
-public record CreateSchemaSqlStatementR(SchemaReference schema, boolean ifNotExists)
-        implements CreateSchemaSqlStatement {
-
-}
+package org.eclipse.daanse.common.jdbc.db.record.schema;
