--- jdbc/db/record/src/main/java/org/eclipse/daanse/common/jdbc/db/record/sql/element/SchemaReferenceR.java
@@ -1,27 +1,0 @@
-/*
-* Copyright (c) 2024 Contributors to the Eclipse Foundation.
-*
-* This program and the accompanying materials are made
-* available under the terms of the Eclipse Public License 2.0
-* which is available at https://www.eclipse.org/legal/epl-2.0/
-*
-* SPDX-License-Identifier: EPL-2.0
-*
-* Contributors:
-*   SmartCity Jena - initial
-*   Stefan Bischof (bipolis.org) - initial
-*/
-package org.eclipse.daanse.common.jdbc.db.record.sql.element;
-
-import java.util.Optional;
-
-import org.eclipse.daanse.common.jdbc.db.api.schema.CatalogReference;
-import org.eclipse.daanse.common.jdbc.db.api.schema.SchemaReference;
-
-public record SchemaReferenceR(Optional<CatalogReference> catalog, String name) implements SchemaReference {
-
-    public SchemaReferenceR(String name) {
-        this(Optional.empty(), name);
-    }
-
-}
