+++ jdbc/loader/csv/src/main/java/org/eclipse/daanse/common/jdbc/loader/csv/impl/CsvDataLoader.java
@@ -44,15 +44,15 @@ import org.eclipse.daanse.common.jdbc.db.api.schema.ColumnReference;
 import org.eclipse.daanse.common.jdbc.db.api.schema.SchemaReference;
 import org.eclipse.daanse.common.jdbc.db.api.schema.TableReference;
 import org.eclipse.daanse.common.jdbc.db.api.sql.InsertSqlStatement;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.ColumnDefinitionR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.ColumnMetaDataR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.ColumnReferenceR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.SchemaReferenceR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.TableReferenceR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.statement.CreateContainerSqlStatementR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.statement.CreateSchemaSqlStatementR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.statement.DropContainerSqlStatementR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.statement.InsertSqlStatementR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.ColumnDefinitionR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.ColumnMetaDataR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.ColumnReferenceR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.SchemaReferenceR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.TableReferenceR;
+import org.eclipse.daanse.common.jdbc.db.record.sql.CreateContainerSqlStatementR;
+import org.eclipse.daanse.common.jdbc.db.record.sql.CreateSchemaSqlStatementR;
+import org.eclipse.daanse.common.jdbc.db.record.sql.DropContainerSqlStatementR;
+import org.eclipse.daanse.common.jdbc.db.record.sql.InsertSqlStatementR;
 import org.eclipse.daanse.common.jdbc.loader.csv.api.Constants;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
