+++ jdbc/loader/csv/src/test/java/org/eclipse/daanse/common/jdbc/loader/csv/CsvDataLoaderTest.java
@@ -37,8 +37,8 @@ import org.eclipse.daanse.common.jdbc.db.api.DatabaseService;
 import org.eclipse.daanse.common.jdbc.db.api.meta.TableDefinition;
 import org.eclipse.daanse.common.jdbc.db.api.schema.ColumnDefinition;
 import org.eclipse.daanse.common.jdbc.db.api.schema.TableReference;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.SchemaReferenceR;
-import org.eclipse.daanse.common.jdbc.db.record.sql.element.TableReferenceR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.SchemaReferenceR;
+import org.eclipse.daanse.common.jdbc.db.record.schema.TableReferenceR;
 import org.eclipse.daanse.common.jdbc.loader.csv.api.Constants;
 import org.h2.jdbcx.JdbcDataSource;
 import org.junit.jupiter.api.AfterEach;
